requirements/base.txt
asgiref==3.2.3
astroid==2.3.3
coverage==5.0.3
Django==3.0.2
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.1
tzlocal==2.0.0
wrapt==1.11.2
